requirements.txt
# Django Framework
Django==4.2.7

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Environment Variables
python-dotenv==1.0.0

# Email (optional, cho production)
django-anymail==10.2  # Hỗ trợ nhiều email service

# Security
django-cors-headers==4.3.1  # CORS nếu có API

# # For Chatbot (Member 4 - Tuần 3)
# nltk==3.8.1

# Testing & Development
coverage==7.3.2  # Test coverage
django-debug-toolbar==4.2.0  # Debug tool (chỉ dùng development)

# Task Queue (optional - cho background jobs)
# celery==5.3.4
# redis==5.0.1

# API (optional - nếu cần REST API)
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7

# Captcha
django-simple-captcha==0.5.21

# Production Server
gunicorn==21.2.0  # WSGI server
whitenoise==6.6.0  # Serve static files

# Monitoring (optional)
# sentry-sdk==1.38.0  # Error tracking
Pillow